+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # CHANGELOG
 
-## master
+## 0.2.26 Feb 5, 2023
 
 - api, api-{augment, base, contract, derive}, rpc-{augment, core, provider}, types, types-{augment, codec, create, known} 9.13.6
-- api, api-{augment, base, contract, derive}, rpc-{augment, core, provider}, types, types-{augment, codec, create, known} 9.13.5
+
 
 ## 0.2.25 Feb 1, 2023